% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_plumber_file.R
\name{build_plumber_file}
\alias{build_plumber_file}
\title{Build the plumber.R file}
\usage{
build_plumber_file(pkg = ".")
}
\arguments{
\item{pkg}{path to package}
}
\description{
Build the plumber.R file
}
