% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pipework.R
\name{create_mariobox}
\alias{create_mariobox}
\title{Create a {mariobox} project to package a {plumber} API}
\usage{
create_mariobox(
  path,
  open = TRUE,
  overwrite = FALSE,
  package_name = basename(path)
)
}
\arguments{
\item{path}{A character string indicating the path (folder) where to create
the package in. The folder name will also be used as the package name.}

\item{open}{A logical. Should the new project be open?}

\item{overwrite}{A logical. Should the already existing project be overwritten ?}

\item{package_name}{Package name to use. By default, {mariobox} uses
\code{basename(path)}. If \code{path == '.'} & \code{package_name} is not explicitly set,
then \code{basename(getwd())} will be used.}
}
\value{
The path to the project. Invisibly
}
\description{
This function will create a prepopulated package
with all the necessary elements to publish a {plumber} API as a package.
}
\examples{
path_pipo <- tempfile(pattern = "pipo")
create_mariobox(
  path = path_pipo,
  open = FALSE
)

}
