% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_error.R
\name{http_error}
\alias{http_error}
\title{Produce HTTP errors}
\usage{
http_error(status = 500L, message = NULL)
}
\arguments{
\item{status}{Integer, HTTP status code for the error (4xx or 5xx).}

\item{message}{Character, the error message (can be a vector).}
}
\value{
A condition object of class \code{"http_error"}.
}
\description{
This function is similar to \code{\link[=stop]{stop()}} and is used to automatically
distinguish server vs. client errors in a Plumber API.
}
\examples{
## R's default error
str(attr(try(stop("Hey, stop!")), "condition"))

## default status code 500 with default error message
str(attr(try(http_error()), "condition"))

## custom status code with default error message
str(attr(try(http_error(501L)), "condition"))

## custom status code for client error
str(attr(try(http_error(400L, "Provide valid email address.")), "condition"))

}
