% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_endpoints.R
\name{add_endpoint}
\alias{add_endpoint}
\alias{manage_endpoints}
\alias{add_delete}
\alias{add_get}
\alias{add_patch}
\alias{add_post}
\alias{add_put}
\alias{remove_endpoint}
\title{Manage endpoints}
\usage{
add_endpoint(name, method = "GET", open = TRUE, pkg = ".")

add_delete(name, open = TRUE, pkg = ".")

add_get(name, open = TRUE, pkg = ".")

add_patch(name, open = TRUE, pkg = ".")

add_post(name, open = TRUE, pkg = ".")

add_put(name, open = TRUE, pkg = ".")

remove_endpoint(name, method, pkg = ".")
}
\arguments{
\item{name}{A character string. The name of the endpoint.}

\item{method}{A character string. The name of the rest method.}

\item{open}{A logical. Should the project be exported or not?}

\item{pkg}{A character string. The path to package root.}
}
\value{
Nothing. Used for its side effect.
}
\description{
Add or remove endpoints in a mariobox project.
}
\examples{
\dontrun{
# Create a new mariobox project
path_pipo <- tempfile(pattern = "pipo")
create_mariobox(
  path = path_pipo,
  open = FALSE
)
# Add an endpoint
add_endpoint(
  name = "allo",
  method = "GET",
  open = FALSE,
  pkg = path_pipo
)
# Remove endpoint
remove_endpoint(
  name = "allo",
  pkg = path_pipo
)
}
}
