% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_error.R
\name{mario_try}
\alias{mario_try}
\title{Try an expression in a Plumber endpoint allowing error recovery}
\usage{
mario_try(res, expr, silent = TRUE, ...)
}
\arguments{
\item{res}{The HTTP response object.}

\item{expr}{An R expression to try.}

\item{silent}{Logical, should the report of error messages be suppressed?}

\item{...}{Other arguments passed to \code{\link[=try]{try()}}.}
}
\value{
The value of the expression if 'expr' is evaluated without error,
but an invisible object of class ‘"try-error"’ containing the
error message, and the error condition as the ‘"condition"’
attribute, if it fails.
}
\description{
This function is similar to \code{\link[=try]{try()}} and is called inside a Plumber API endpoint
definition. The function needs access to Plumbers response object.
Use \code{\link[=http_error]{http_error()}} to automate the handling of HTTP status codes
and differentiate between server vs. client errors.
}
