% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_endpoints.R
\name{pw_add_endpoint}
\alias{pw_add_endpoint}
\alias{manage_endpoints}
\alias{pw_remove_endpoint}
\title{Manage endpoints}
\usage{
pw_add_endpoint(name, methods = "GET", open = TRUE, pkg = ".")

pw_remove_endpoint(name, pkg = ".")
}
\arguments{
\item{name}{A character string. The name of the endpoint.}

\item{methods}{A character string. The name of the rest methods.}

\item{open}{A logical. Should the project be exported or not?}

\item{pkg}{A character string. The path to package root.}
}
\value{
Nothing. Used for its side effect.
}
\description{
Add or remove endpoints in a pipework project.
}
\examples{
\dontrun{
# Create a new pipework project
path_pipo <- tempfile(pattern = "pipo")
create_pipework(
  path = path_pipo,
  open = FALSE
)
# Add an endpoint
pw_add_endpoint(
  name = "allo",
  methods = "GET",
  open = FALSE,
  pkg = path_pipo
)
# Remove endpoint
pw_remove_endpoint(
  name = "allo",
  pkg = path_pipo
)
}
}
